Move postContent.ts from lib/ -> client/
