+++ packages/api/src/client/postsApi.ts
@@ -2,10 +2,6 @@ import { da, render } from '@urbit/aura';
 
 import type { Poke } from '../http-api';
 import { createDevLogger } from '../lib/logger';
-import {
-  PlaintextPreviewConfig,
-  getTextContent as getTextPreview,
-} from '../lib/postContent';
 import { IMAGE_URL_REGEX } from '../lib/utils';
 import type * as db from '../types/models';
 import { ContentReference } from '../types/references';
@@ -46,6 +42,10 @@ import {
 } from './apiUtils';
 import { channelAction } from './channelsApi';
 import { multiDmAction } from './chatApi';
+import {
+  PlaintextPreviewConfig,
+  getTextContent as getTextPreview,
+} from './postContent';
 import { poke, scry, subscribeOnce } from './urbit';
 
 const logger = createDevLogger('postsApi', false);
